+++ core-ui/RELEASE_NOTES.md
@@ -21,6 +21,7 @@
 * ANDDEP-1048 Fixing wrong docs links and docs structure
 * ANDDEP-1138 Added supporting View Binding
 * fix getting root view, return root view instead DecorView, because view binding can't init by DecorView
+* Added android 11 support for PermissionManager (one-time permission handling)
 ## 0.4.0
 ##### Core-ui
 * Added the ability to set text for the buttons of the standard dialog for switching to settings.